Refactor #1
