+++ enterprise/frontend/src/metabase-enterprise/data_apps/admin/DataAppSkillsSection.module.css
@@ -1,0 +1,23 @@
+.command {
+  padding: 1rem;
+  padding-right: 3rem;
+  border: 1px solid var(--mb-color-border);
+  border-radius: var(--mantine-radius-md);
+  background-color: var(--mb-color-background-secondary);
+  overflow-x: auto;
+}
+
+.commandText {
+  margin: 0;
+  font-family: var(--mb-default-monospace-font-family, monospace);
+  font-size: 0.75rem;
+  line-height: 1.6;
+  color: var(--mb-color-text-primary);
+  white-space: pre;
+}
+
+.copyButton {
+  position: absolute;
+  top: 0.5rem;
+  right: 0.5rem;
+}
