+++ enterprise/frontend/src/metabase-enterprise/data_apps/admin/ManageDataAppsPage.tsx
@@ -5,12 +5,9 @@ import {
   SettingsSection,
 } from "metabase/admin/components/SettingsSection";
 import { Link } from "metabase/common/components/Link";
-import { useSetting } from "metabase/common/hooks";
 import {
-  ActionIcon,
   Box,
   Button,
-  CopyButton,
   Flex,
   Group,
   Icon,
@@ -18,35 +15,17 @@ import {
   Stack,
   Text,
   Title,
-  Tooltip,
 } from "metabase/ui";
-import {
-  isLocalOrSnapshotVersion,
-  versionToNumericComponents,
-} from "metabase/utils/version";
 import {
   useGetDataAppRepoStatusQuery,
   useListDataAppsQuery,
 } from "metabase-enterprise/api";
 
 import { DataAppListItem } from "./DataAppListItem";
+import { DataAppSkillsSection } from "./DataAppSkillsSection";
 import S from "./ManageDataAppsPage.module.css";
 
 const REMOTE_SYNC_SETTINGS_PATH = "/admin/settings/remote-sync";
-
-const REPOSITORY_NAME = "metabase/metabase";
-const MAIN_BRANCH_NAME = "master";
-const RELEASE_BRANCH_PREFIX = "release";
-const PUBLIC_SKILLS_PATH = "/skills";
-
-// The data-app skills to install. `skills add metabase/metabase` alone would
-// discover *every* skill in the repo, so each is selected explicitly.
-const DATA_APP_SKILLS = [
-  "metabase-data-app-setup",
-  "metabase-data-app-routing",
-  "metabase-data-app-actions",
-  "metabase-data-app-semantic-layer",
-];
 
 // Keep the sync status fresh while the page is open: the backend polls the repo
 // on its own schedule, so re-fetch the list periodically to reflect those syncs.
@@ -63,7 +42,7 @@ const POLL_OPTS = {
   refetchOnMountOrArgChange: true,
 } as const;
 
-export function ManageDataAppsPage() {
+export const ManageDataAppsPage = () => {
   const { data: status, isLoading: isStatusLoading } =
     useGetDataAppRepoStatusQuery(undefined, POLL_OPTS);
   const { data: apps, isLoading: isAppsLoading } = useListDataAppsQuery(
@@ -72,28 +51,6 @@ export function ManageDataAppsPage() {
   );
 
   const isConfigured = status?.configured ?? false;
-
-  // Pin the data-app skills (and the template bundled inside `metabase-data-app-setup`)
-  // to the branch matching this instance: `release-x.<major>.x`, or `master` for
-  // local/dev builds that have no release branch.
-  const { tag } = useSetting("version");
-
-  const majorVersion = tag ? versionToNumericComponents(tag)?.[1] : undefined;
-
-  // eslint-disable-next-line metabase/no-literal-metabase-strings -- Admin UI string
-  const skillsMainText = t`Install Metabase Data App skills in your project, then ask your AI agent to create a data app.`;
-  const skillBranch =
-    tag && !isLocalOrSnapshotVersion(tag) && majorVersion != null
-      ? `${RELEASE_BRANCH_PREFIX}-x.${majorVersion}.x`
-      : MAIN_BRANCH_NAME;
-  const skillCommandBase = `npx skills add ${REPOSITORY_NAME}${PUBLIC_SKILLS_PATH}#${skillBranch}`;
-  const skillSelectors = DATA_APP_SKILLS.map((skill) => `--skill ${skill}`);
-  // One line to copy (a runnable command); shown wrapped across lines for
-  // readability.
-  const installSkillCommand = [skillCommandBase, ...skillSelectors].join(" ");
-  const installSkillCommandDisplay = [skillCommandBase, ...skillSelectors].join(
-    "\n",
-  );
 
   return (
     <SettingsPageWrapper>
@@ -127,6 +84,7 @@ export function ManageDataAppsPage() {
                 bg="background-secondary"
                 bd="1px solid var(--mb-color-border)"
                 bdrs="md"
+                visibleFrom="sm"
               >
                 <Icon
                   name="git_branch"
@@ -142,6 +100,7 @@ export function ManageDataAppsPage() {
                   {isConfigured ? status?.url : t`No repository connected`}
                 </Text>
               </Group>
+
               <Button
                 component={Link}
                 to={REMOTE_SYNC_SETTINGS_PATH}
@@ -152,33 +111,7 @@ export function ManageDataAppsPage() {
             </Group>
           </Stack>
 
-          <Stack gap="sm">
-            <Title order={3}>{t`AI skills`}</Title>
-
-            <Text>{skillsMainText}</Text>
-
-            <Box className={S.command} pos="relative">
-              <CopyButton value={installSkillCommand} timeout={2000}>
-                {({ copied, copy }) => (
-                  <Tooltip label={copied ? t`Copied!` : t`Copy`}>
-                    <ActionIcon
-                      className={S.copyButton}
-                      variant="subtle"
-                      c="text-secondary"
-                      aria-label={t`Copy`}
-                      data-testid="copy-button"
-                      onClick={copy}
-                    >
-                      <Icon name="copy" />
-                    </ActionIcon>
-                  </Tooltip>
-                )}
-              </CopyButton>
-              <Text component="pre" className={S.commandText}>
-                {installSkillCommandDisplay}
-              </Text>
-            </Box>
-          </Stack>
+          <DataAppSkillsSection />
         </SettingsSection>
       </Stack>
 
@@ -250,4 +183,4 @@ export function ManageDataAppsPage() {
       )}
     </SettingsPageWrapper>
   );
-}
+};
